update project angular

+++ module5/project/angular-project/src/app/products/product-list/product-list.component.ts
@@ -1,8 +1,12 @@
+import { CreateProductComponent } from './../create-product/create-product.component';
+import { ProductDetailComponent } from './../product-detail/product-detail.component';
 import { IProduct } from './../model/product';
 import { ProductService } from './../service/product.service';
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Location } from '@angular/common';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDeleteComponent } from '../confirm-delete/confirm-delete.component';
 @Component({
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
@@ -10,10 +14,11 @@ import { Location } from '@angular/common';
 })
 export class ProductListComponent implements OnInit,OnChanges {
   products : IProduct[];
-  search : String ='';
+  search : string ='';
   constructor(private productService : ProductService,
     private activatedRoute : ActivatedRoute,
-    private router : Router) { }
+    private router : Router,
+    public dialog: MatDialog) { }
   ngOnChanges(changes: SimpleChanges): void {
   }
   ngOnInit(): void {
@@ -23,15 +28,34 @@ export class ProductListComponent implements OnInit,OnChanges {
     this.initDataListProduct();
   }
   initDataListProduct() {
-    if(!this.search) {
       this.products =this.productService.getAllProduct()
-    } else{
-      this.products=this.productService.getBySearch(this.search);
-    }
-
+  }
+  handleCreate() {
+    const dialogRef = this.dialog.open(CreateProductComponent, {
+      width : '350px' ,
+      height : 'auto',
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      this.productService.createProduct(result.value);
+    });
   }
   handleDelete(id : number) {
-    console.log(id);
-    this.productService.deleteProduct(id);
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent, {
+      width : '300px' ,
+      data: { product : this.productService.getById(id) }
+    });
+    dialogRef.afterClosed().subscribe(result =>  {
+      if(result) {this.productService.deleteProduct(id)};
+    });
+  }
+  handleView(id : number) {
+    const dialogRef = this.dialog.open(ProductDetailComponent, {
+      width : '350px' ,
+      height : 'auto',
+      data: { product : this.productService.getById(id) }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(`Dialog result: ${result}`);
+    });
   }
 }
